opengl/requirements.txt
glfw==2.6.3
Pillow==10.1.0
PyOpenGL==3.1.7